HBASE-9248 Place holders for tags in 0.96 to accommodate tags in 0.98
git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1516614 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-common/src/main/java/org/apache/hadoop/hbase/Cell.java
@@ -181,6 +181,6 @@ public interface Cell {
   /**
    * @return the total length of the tags in the Cell.
    */
-  int getTagsLength();
+  short getTagsLength();
 
 }
